requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
databases==0.8.0
aiosqlite==0.19.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
aiofiles==23.2.1
